+++ core/settings.py
@@ -389,6 +389,9 @@ QFDMO_GOOGLE_SEARCH_CONSOLE = "google9dfbbc61adbe3888.html"
 SECURE_REFERRER_POLICY = "strict-origin-when-cross-origin"
 ASSISTANT = {
     "MATOMO_ID": decouple.config("ASSISTANT_MATOMO_ID", default=82, cast=int),
+    "HOSTS": decouple.config("ASSISTANT_HOSTS", default="localhost", cast=str).split(
+        ","
+    ),
     "POSTHOG_KEY": decouple.config(
         "ASSISTANT_POSTHOG_KEY",
         default="phc_fSfhoWDOUxZdKWty16Z3XfRiAoWd1qdJK0N0z9kQHJr",  # pragma: allowlist secret  # noqa: E501
