+++ core/urls.py
@@ -25,6 +25,7 @@ from django.views.generic import TemplateView
 from qfdmo.models.acteur import ActeurStatus, DisplayedActeur
 
 from .api import api
+from .views import direct_access
 
 info_dict = {
     "queryset": DisplayedActeur.objects.filter(statut=ActeurStatus.ACTIF).order_by(
@@ -60,6 +61,7 @@ urlpatterns = [
     path("dsfr/", include(("dsfr_hacks.urls", "dsfr_hacks"), namespace="dsfr_hacks")),
     path("", include(("qfdmo.urls", "qfdmo"), namespace="qfdmo")),
     path("", include(("qfdmd.urls", "qfdmd"), namespace="qfdmd")),
+    path("", direct_access, name="home"),
     path("data/", include(("data.urls", "data"), namespace="data")),
     path("docs/", TemplateView.as_view(template_name="techdocs.html"), name="techdocs"),
 ]
