+++ core/views.py
@@ -1,8 +1,11 @@
 import mimetypes
 
+from django.conf import settings
 from django.contrib.auth.mixins import LoginRequiredMixin
 from django.contrib.staticfiles import finders
 from django.http import HttpResponse
+from django.shortcuts import redirect
+from django.urls.base import reverse
 
 
 def static_file_content_from(path):
@@ -19,6 +22,29 @@ def static_file_content_from(path):
     with open(file_path, "r") as file:
         file_content = file.read()
         return HttpResponse(file_content, content_type=content_type)
+
+
+def direct_access(request):
+    from qfdmd.views import HomeView as Assistant  # avoid circular dependency
+
+    get_params = request.GET.copy()
+
+    if "carte" in request.GET:
+        # Order matters, this should be before iframe because iframe and carte
+        # parameters can coexist
+        del get_params["carte"]
+        try:
+            del get_params["iframe"]
+        except KeyError:
+            pass
+        params = get_params.urlencode()
+        parts = [reverse("qfdmo:carte"), "?" if params else "", params]
+        return redirect("".join(parts))
+
+    if request.META["HTTP_HOST"] in settings.ASSISTANT["HOSTS"]:
+        return Assistant.as_view()(request)
+
+    return redirect("https://longuevieauxobjets.ademe.fr/lacarte", permanent=True)
 
 
 class IsStaffMixin(LoginRequiredMixin):
