+++ templates/components/header/header.html
@@ -4,7 +4,7 @@
         <a
             class="qf-bg-none qf-flex qf-gap-2w"
             data-testid="header-logo-link"
-            href="{% url 'qfdmd:home' %}"
+            href="{% url 'home' %}"
         >
             <img
                 src="{% static 'assistant/bloc-marque.svg' %}"
